refactor: remove report script bridges
