+++ docs/steering/structure.md
@@ -98,8 +98,8 @@ the manifest schema, frozen-artifact hashes, and implementation references.
 ### Transitional Script Dependencies
 
 First-party packages may not add dependencies on `scripts/`. The architecture guard carries
-three exact temporary execution bridges from the tech-area, weekly-awards, and
-phase-transition report jobs to their script entry points. These are migration bridges, not a
+one exact temporary execution bridge from the tech-area report job to its script entry point.
+This is a migration bridge, not a
 fifth epistemic tier or an implicit promotion of those scripts. They are limited to named
 compatibility wrappers, must not be used by an evidence-tier artifact, and are removed when
 the implementations move behind package APIs. The CLI modules can then remain as entry points.
